+++ samples/snippets/documentSigners.js
@@ -1,0 +1,32 @@
+'use strict';
+
+const signnow = require('@signnow/api-client')({
+  credentials: 'BASE64_ENCODED_CLIENT_CREDENTIALS',
+  production: true, // if false then uses eval server
+});
+const documentSigners = signnow.document.signers;
+
+const id = 'DOCUMENT_ID_GOES_HERE';
+const token = 'YOUR_ACCESS_TOKEN';
+
+/**
+ * @param {Array} res - emails of document signers
+ */
+const handleResponse = res => {
+  console.log(res);
+};
+
+const handleError = err => {
+  console.error(err);
+};
+
+documentSigners({
+  id,
+  token,
+}, (err, res) => {
+  if (err) {
+    handleError(err);
+  } else {
+    handleResponse(res);
+  }
+});